music recommendations and miscellaneous fixes

+++ src/main/java/tk/ardentbot/Core/WebServer/Models/Status.java
@@ -7,9 +7,9 @@ public class Status {
     public long receivedMessages;
     public long receivedCommands;
     public long uptime;
-    public int loaded_commands;
-    public int guilds;
-    public int users;
+    public long loaded_commands;
+    public long guilds;
+    public long users;
 
     public Status(long receivedMessages, long receivedCommands, long uptime, long loaded_commands, long guilds, long
             users) {
